Address Windows build warning: InterlockedCompareExchange() expects a long*, not a uint32_t*.

+++ src/xpdev/threadwrap.h
@@ -124,7 +124,7 @@ DLLEXPORT int DLLCALL pthread_mutex_destroy(pthread_mutex_t*);
 
 // A structure in case we need to add an event or something...
 typedef struct {
-	uint32_t	state;
+	long	state;
 } pthread_once_t;
 
 #define PTHREAD_ONCE_INIT	{0};
